infra/lambda/dspy-insights/requirements.txt
dspy-ai>=2.4.0
boto3>=1.34.0

